src/functions/postgresql_connection_test/requirements.txt
psycopg2-binary
requests
azure-identity
